+++ examples/live/binance/binance_futures_testnet_exec_tester.py
@@ -130,6 +130,7 @@ strat_config = ExecTesterConfig(
     instrument_id=InstrumentId.from_str("ETHUSDT-PERP.BINANCE"),
     external_order_claims=[InstrumentId.from_str("ETHUSDT-PERP.BINANCE")],
     order_qty=order_qty,
+    order_params={"price_match": "QUEUE_5"},
     # open_position_on_start_qty=order_qty,
     # tob_offset_ticks=0,
     # use_batch_cancel_on_stop=True,